BSP-145: Replaced individual controllers with single BulkScanController
